Improve user functionality and display country flag
Searches can now be performed using the return key as well as the search button. A request is made to the CountryFlags API on search submission, using the country code returned from the Open Weather Map API GET request, and is used to display the flag of the country in question at the top right of the page.